Lots of changes.
Add ability to turn APRS publication on or off.  This is done in the settings
dialog.

Clean up code for generating APRS --- move it off into routines that check
whether we've disabled APRS or not.

Make ML fix more robust, checking for invalid returns and using multiple
techniques for geting at it.

Add Stansfield computation, and display of Stansfield confidence intervals.