Add capability to edit reports after they are created.  This took for-frickin' ever to commit, because it took me forever to find time to move forward on it.
Lotta changes here.  Had to revamp a lot of how the qDFProjReport and the
collection work internally, and had to add all sorts of accessors to the
report and collection.  Had to create an alternate constructor for the
report dialog.  Had to add a function to the latLonCoordinateWidget so that
one could pre-set the contents.  Eliminated a valgrind-detected error
where we were testing lat/lon before setting them in the isValid method.

Editing of reports now works.