Fix dumb mistake in computation of stansfield error ellipses.
The equation of the ellipse is:

   X^2/a^2 + Y^2/b^2 = -2ln(1-P)

so -2ln(1-P) is supposed to be 'rho^2', but I was using it as rho.  That is:

X = a*rho*cos(theta)
Y = b*rho*sin(theta)

is the polar coordinate form of the ellipse.  Only when using
rho^2=-2ln(1-P) does this actually solve the original ellipse
equation.  The result was that the ellipses I was drawing were way too
large.


In CoordSysBuilder, fix "zone = " vs. "zone=" thing for UTM coordinate
systems.  This is utterly inconsequential in the current version,
which never uses that code.