Start working on "display manager" idea.
First step is to create an abstract "qDFDisplayInterface" class and
implement its methods in the MainWindow for the textual display there.

This commit does that.

Next step:  do the same for the APRS display.
After that: migrate calls to both display interfaces into a display manager
that loops over all the display interfaces it owns.
Phase 3: profit.