Rethink the clear/initialize/close stuff a little, and use them properly.
In pondering how I'll do the KML display, I realize that I don't want
qDF clearing out the KML file when closing the DF solution file.  We could
easily want to save that KML file as a record of the DF search.

Also rename initialize/clear/close to intializeDisplay etc. to keep them
from interfering with stuff like slots.

So here's the skinny:

  initializeDisplay() is called when beginning a new DF problem.
  clearDisplay() is called whenever we want to wipe out displays
  closeDisplay() is called whenever we close a DF problem.

In the MainWindow and aprsDisplay classes, closeDisplay and
initializeDisplay just call clearDisplay(), since that's the desired
behavior of those display methods.  When I write KMLDisplay, it'll be
different.