Knock off another "TODO" item.
Implemented a "Display Manager" into which one adds all the display
interfaces one wants.  When the main code wants to display anything,
it calls the display manager's display function, not the interfaces'
functions.  The display manager takes care of farming out the calls.

This simplifies the top level code and makes it easier to encapsulate
interface-specific junk in the code for just that interface.

More refinement:  closeDisplays is called when we clear the report
collection.  initializeDisplays() is called when we start entering data
for a new collection.

But enableDisplay also calls the initializeDisplay for that interface if
it was not already enabled, so it should NOT clear the display, just set up
for display.

This is in anticipation of the KML output method, where it is not exactly
clear yet how we handle start-up and close.  That's the next line of
development.