Make qDF stop sending sequence numbers on its multiline objects.
When I implemented the APRS multiline code in qDF, it was unclear to
me that the APRS Multiline spec allowed the sequence number to be
optional.  The sequence number is always interpreted as an NWS alert
ID by APRS clients, and having qDF put something there was always an
issue --- there was no WS alert associated with the sequence number.

Turns out that the sequence number is optional, so I won't transmit one
anyore.  Leaving it out makes room for more coordinate pairs, as well.