Get qDF working again
This commit updates qDF to use Qt6.

It does make use of the deprecated function QString::asprintf in place
of the sprintf QString member function that no longer exists.  A
future fix might be to remove the asprintf usage and replace with a
QTextStream, but the very reason I'm using sprintf is to force the
exact formatting of strings for APRS, and it's easy to do with sprintf
and not with streams (lots of setting of precision, width, padding
char, etc.).

Also fix the broken computation of Stansfield fix, which wasn't being
given an appropriate starting guess and so was failing silently
because we were just catching the exception and quietly setting the
fix to invalid (now we print out that the exception happened if it
does fail).

Tested out with old data files and it regenerated all the displays I
expected.

Closes #4